Split integration:set into add and update
