+++ src/Application.php
@@ -110,9 +110,10 @@ class Application extends ParentApplication {
         $commands[] = new Command\EnvironmentVariableDeleteCommand();
         $commands[] = new Command\EnvironmentVariableGetCommand();
         $commands[] = new Command\EnvironmentVariableSetCommand();
-        $commands[] = new Command\IntegrationSetCommand();
+        $commands[] = new Command\IntegrationAddCommand();
         $commands[] = new Command\IntegrationDeleteCommand();
         $commands[] = new Command\IntegrationGetCommand();
+        $commands[] = new Command\IntegrationUpdateCommand();
         $commands[] = new Command\ProjectBuildCommand();
         $commands[] = new Command\ProjectCleanCommand();
         $commands[] = new Command\ProjectDrushAliasesCommand();
